refactor(flow): node-type registry — first-class self-describing nodes
Foundation for growing the node catalog (transform / http / human-review
/ integrations) without touching the core. Before: a node type was an
implicit convention scattered across 6 places (graph const + knownTypes +
portsFor special-case + Triggers string-prefix + a 250-line execute
switch + frontend hardcoding). Adding one meant editing all of them.

Now every node type is one self-describing NodeSpec — Kind (trigger/
action/control/data) / Category / Ports / Execute — in a registry. The
engine, graph validation, port checks, and trigger detection all DERIVE
from it:
- execute() is pure dispatch: LookupNode(type).Execute(...)
- knownTypes map → LookupNode != nil
- portsFor special-case → spec.Ports
- Triggers()/edge-into-trigger string-prefix → spec.Kind == KindTrigger

The 8 built-ins migrate verbatim (behaviour unchanged; all existing
tests green). Adding a node type is now RegisterNode(one spec) — proven
by TestNodeRegistryExtensibility (a custom node runs end-to-end with zero
core edits). Execute/RunSingle signatures unchanged. No new deps.
