+++ internal/manager/biz/flow/nodes.go
@@ -68,161 +68,15 @@ const (
 	llmNodeTimeout     = 3 * time.Minute
 )
 
-// execute runs one node. cfg is the node's config AFTER expression
-// resolution (every string leaf already substituted).
+// execute runs one node by dispatching to its registered NodeSpec. cfg is
+// the node's config AFTER expression resolution. The per-type behaviour
+// lives in noderegistry.go — this is pure dispatch, no switch.
 func (x Executors) execute(ctx context.Context, node GraphNode, cfg map[string]any, rc *RunContext) (NodeResult, error) {
-	switch node.Type {
-	case NodeTriggerManual, NodeTriggerAlert, NodeTriggerCron:
-		// Every trigger's "output" is the trigger payload itself so
-		// downstream can use either {{trigger.x}} or {{nodes.<id>.output.x}}.
-		// The payload differs by source (manual input / incident context /
-		// cron fire time) but the node behaviour is identical.
-		return NodeResult{Output: anyMap(rc.Trigger), Port: PortNext}, nil
-
-	case NodeAgent:
-		if x.Agent == nil {
-			return NodeResult{}, fmt.Errorf("agent node: runner not wired")
-		}
-		persona, _ := cfg["persona"].(string)
-		if persona == "" {
-			persona = "default"
-		}
-		instruction, _ := cfg["instruction"].(string)
-		if strings.TrimSpace(instruction) == "" {
-			return NodeResult{}, fmt.Errorf("agent node: instruction is empty")
-		}
-		// Structured gateway (HLD-016): with an output_schema the agent
-		// is instructed to answer ONLY the JSON object; we parse it into
-		// output.structured so deterministic downstream (condition /
-		// tool args) can reference fields. Without one the answer is
-		// free text — consumable by agents / notify / humans only.
-		schema, hasSchema := cfg["output_schema"].(map[string]any)
-		prompt := instruction
-		if hasSchema {
-			sb, _ := json.Marshal(schema)
-			prompt = instruction + "\n\nReturn ONLY a single JSON object matching this JSON Schema (no prose, no code fence):\n" + string(sb)
-		}
-		actx, cancel := context.WithTimeout(ctx, agentNodeTimeout)
-		defer cancel()
-		answer, err := x.Agent.RunAgent(actx, persona, prompt)
-		if err != nil {
-			return NodeResult{}, fmt.Errorf("agent node: %w", err)
-		}
-		out := map[string]any{"answer": answer}
-		if hasSchema {
-			structured, perr := parseLooseJSON(answer)
-			if perr != nil {
-				return NodeResult{}, fmt.Errorf("agent node: output_schema declared but answer is not JSON: %w", perr)
-			}
-			out["structured"] = structured
-		}
-		return NodeResult{Output: out, Port: PortNext}, nil
-
-	case NodeLLM:
-		if x.LLM == nil {
-			return NodeResult{}, fmt.Errorf("llm node: runner not wired")
-		}
-		prompt, _ := cfg["prompt"].(string)
-		if strings.TrimSpace(prompt) == "" {
-			return NodeResult{}, fmt.Errorf("llm node: prompt is empty")
-		}
-		system, _ := cfg["system"].(string)
-		// Same structured gateway as the agent node: an output_schema
-		// turns the free-text answer into output.structured so condition
-		// / tool nodes can reference fields.
-		schema, hasSchema := cfg["output_schema"].(map[string]any)
-		if hasSchema {
-			sb, _ := json.Marshal(schema)
-			prompt = prompt + "\n\nReturn ONLY a single JSON object matching this JSON Schema (no prose, no code fence):\n" + string(sb)
-		}
-		lctx, cancel := context.WithTimeout(ctx, llmNodeTimeout)
-		defer cancel()
-		answer, err := x.LLM.RunLLM(lctx, system, prompt)
-		if err != nil {
-			return NodeResult{}, fmt.Errorf("llm node: %w", err)
-		}
-		out := map[string]any{"answer": answer}
-		if hasSchema {
-			structured, perr := parseLooseJSON(answer)
-			if perr != nil {
-				return NodeResult{}, fmt.Errorf("llm node: output_schema declared but answer is not JSON: %w", perr)
-			}
-			out["structured"] = structured
-		}
-		return NodeResult{Output: out, Port: PortNext}, nil
-
-	case NodeTool:
-		if x.Tools == nil {
-			return NodeResult{}, fmt.Errorf("tool node: invoker not wired")
-		}
-		name, _ := cfg["tool"].(string)
-		if name == "" {
-			return NodeResult{}, fmt.Errorf("tool node: missing tool name")
-		}
-		args, _ := cfg["args"].(map[string]any)
-		ab, err := json.Marshal(args)
-		if err != nil {
-			return NodeResult{}, fmt.Errorf("tool node: args: %w", err)
-		}
-		tctx, cancel := context.WithTimeout(ctx, defaultNodeTimeout)
-		defer cancel()
-		res, err := x.Tools.InvokeTool(tctx, name, ab)
-		if err != nil {
-			return NodeResult{}, fmt.Errorf("tool %s: %w", name, err)
-		}
-		var out any
-		if len(res) > 0 {
-			if uerr := json.Unmarshal(res, &out); uerr != nil {
-				out = string(res) // non-JSON tool output stays a string
-			}
-		}
-		return NodeResult{Output: map[string]any{"result": out}, Port: PortNext}, nil
-
-	case NodeCondition:
-		expr, _ := cfg["expr"].(string)
-		if strings.TrimSpace(expr) == "" {
-			return NodeResult{}, fmt.Errorf("condition node: missing expr")
-		}
-		ok, err := rc.EvalCondition(expr)
-		if err != nil {
-			return NodeResult{}, err
-		}
-		port := PortFalse
-		if ok {
-			port = PortTrue
-		}
-		return NodeResult{Output: map[string]any{"result": ok}, Port: port}, nil
-
-	case NodeNotify:
-		if x.Notify == nil {
-			return NodeResult{}, fmt.Errorf("notify node: notifier not wired")
-		}
-		title, _ := cfg["title"].(string)
-		message, _ := cfg["message"].(string)
-		if strings.TrimSpace(message) == "" {
-			return NodeResult{}, fmt.Errorf("notify node: message is empty")
-		}
-		ids := toUint64s(cfg["channel_ids"])
-		if len(ids) == 0 {
-			return NodeResult{}, fmt.Errorf("notify node: no channel_ids")
-		}
-		nctx, cancel := context.WithTimeout(ctx, defaultNodeTimeout)
-		defer cancel()
-		if err := x.Notify.Notify(nctx, ids, title, message); err != nil {
-			return NodeResult{}, fmt.Errorf("notify node: %w", err)
-		}
-		return NodeResult{Output: map[string]any{"sent": true, "channels": len(ids)}, Port: PortNext}, nil
-
-	case NodeSet:
-		name, _ := cfg["name"].(string)
-		if name == "" {
-			return NodeResult{}, fmt.Errorf("set node: missing var name")
-		}
-		val := cfg["value"]
-		rc.Vars[name] = val
-		return NodeResult{Output: map[string]any{"name": name, "value": val}, Port: PortNext}, nil
+	spec := LookupNode(node.Type)
+	if spec == nil || spec.Execute == nil {
+		return NodeResult{}, fmt.Errorf("unknown node type %q", node.Type)
 	}
-	return NodeResult{}, fmt.Errorf("unknown node type %q", node.Type)
+	return spec.Execute(ctx, x, cfg, rc)
 }
 
 // parseLooseJSON accepts a raw JSON object possibly wrapped in a code
